Fix app and web

+++ web.py
@@ -24,41 +24,38 @@ response_queue_url: str = (
 
 
 @application.route("/")
-def hello_world():
-    return "Hello World! Go to upload to upload your images!"
+def hello_world() -> str:
+    return "Hello World! Go to the upload page to upload your images!"
 
 
 @application.route("/upload", methods=["POST", "GET"])  # type: ignore
 def upload() -> None | str:
     if request.method == "POST":
-        file3 = request.files.getlist("myfile")
-        for file2 in file3:
-            if file2.filename is None:
-                continue
-            filename = secure_filename(file2.filename)
-            path = os.path.join(application.config["UPLOAD_FOLDER"], filename)
-            file2.save(path)
-            with open(path, "rb") as image2string:
-                bytes = base64.b64encode(image2string.read())
-                sqs.send_message(
-                    QueueUrl=request_queue_url,
-                    MessageAttributes={
-                        "ImageName": {"DataType": "String", "StringValue": filename}
-                    },
-                    MessageBody=bytes.decode("ascii"),
-                )
-                print(filename)
+        file = request.files.getlist("myfile")[0]
+        if file.filename is None:
+            return ""
+        filename = secure_filename(file.filename)
+        path = os.path.join(application.config["UPLOAD_FOLDER"], filename)
+        file.save(path)
+        with open(path, "rb") as image2string:
+            bytes = base64.b64encode(image2string.read())
+        sqs.send_message(
+            QueueUrl=request_queue_url,
+            MessageAttributes={
+                "ImageName": {"DataType": "String", "StringValue": filename}
+            },
+            MessageBody=bytes.decode("ascii"),
+        )
 
         while True:
             try:
-                print("Checking for response...")
                 queue_attr = sqs.get_queue_attributes(
                     QueueUrl=response_queue_url, AttributeNames=["All"]
                 )
                 num_visible_msg = int(
                     queue_attr["Attributes"]["ApproximateNumberOfMessages"]
                 )
-                print("Visible msg:", num_visible_msg)
+                print(f"Number of visible messages: {num_visible_msg}")
                 if num_visible_msg > 0:
                     msg = sqs.receive_message(
                         QueueUrl=response_queue_url,
@@ -67,20 +64,21 @@ def upload() -> None | str:
                     )
                     body = msg["Messages"][0]["Body"]
                     filename, result = body.split(",")
-
+                    print(f"Received `{body}`")
+                    if filename != file.filename:
+                        continue
                     sqs.delete_message(
                         QueueUrl=response_queue_url,
                         ReceiptHandle=msg["Messages"][0]["ReceiptHandle"],
                     )
-                    return f"{result} for file {filename}"
+                    print("Returning response")
+                    return f"Result for file '{filename}': {result}"
                 else:
                     time.sleep(random.random() * 3 + 1)
                     continue
-
             except Exception as e:
                 print("Exception Occured...")
                 print(e)
-
                 time.sleep(random.random() * 3 + 1)
